[agent-180] Fix the dictionary lookup
Error appears to be being thrown in a race condition between recieving
the list of active processes and then querying it for info.
Added a trap for it occuring, and continuing if the process has gone
away in the mean time.

This fix is required for 1.4.3 ([agent-202]) as currently the process
lookup fails entirely.